+++ packages/widgets-manager/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+### Release [1.0.0] - 2019-06-04
+- 'relative-to-position' position strategy changes format
+- **BREAKING CHANGES**: Format for relative to postion changes from:
+  - { offset: Position } to { offsetX: offsetX, offsetY: offsetY }
+
 ## Release [0.7.0] - 2018-12-12
 ### Release
 - Suport new type lt-webrtc